tlqkf

+++ Assets/Script/Enemy/First_Enemy.cs
@@ -42,7 +42,7 @@ public class First_Enemy : Enemy
         animator.SetBool("attack", false);
     }
 
-    public override void GetDamage(int damage)
+    public override void GetDamage(float damage)
     {
         hp -= damage;
     }
